修改后版本

+++ src/main/java/com/ctbu/controller/EmployeeController.java
@@ -4,10 +4,14 @@ import com.ctbu.entity.Employee;
 import com.ctbu.result.Result;
 import com.ctbu.result.ResultEnum;
 import com.ctbu.service.EmployeeService;
+import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Map;
 
 /**
  * @Description 员工controller层
@@ -15,22 +19,13 @@ import java.util.List;
  */
 @RestController
 @RequestMapping("/employee")
+@Api(description = "员工信息管理")
 public class EmployeeController {
     @Autowired
     EmployeeService employeeService;
-
-    @GetMapping("/listemployee")
-    public Result listeEmployees() {
-        List<Employee> res = employeeService.listEmployees();
-        if (res != null) {
-            return Result.success(res);
-        } else {
-            return Result.error(ResultEnum.DATA_IS_NULL.getCode(), ResultEnum.DATA_IS_NULL.getMsg());
-        }
-    }
-
+    @ApiOperation(value = "添加员工信息")
     @PostMapping("/insertemployee")
-    public Result insertEmployee(@RequestBody Employee employee) {
+    public Result insertEmployee(@Validated Employee employee) {
         try {
             employeeService.insertEmployee(employee);
             return Result.success();
@@ -40,7 +35,8 @@ public class EmployeeController {
     }
 
     @PutMapping("/updateemployee")
-    public Result updateEmployee(@RequestBody Employee employee) {
+    @ApiOperation(value = "更新员工信息")
+    public Result updateEmployee(@Validated Employee employee) {
         try {
             employeeService.updateEmployee(employee);
             return Result.success();
@@ -49,26 +45,45 @@ public class EmployeeController {
         }
     }
 
-    @GetMapping("/getEmployeeById")
-    public Result getEmployeeById(String id) {
-        Employee res = employeeService.getEmployeeById(id);
-        if (res != null) {
+
+    @DeleteMapping("/deleteemployee")
+    @ApiOperation(value = "通过员工编号删除员工")
+    public Result deleteEmployee(@RequestParam Integer id) {
+        try {
+            Map<String,Object>  res = employeeService.getEmployeeById(id);
+            if (res != null&&!res.isEmpty()) {
+                employeeService.deleteEmployee(id);
+                return Result.success();
+            }
+            return Result.error(ResultEnum.DATA_IS_NULL.getCode(), ResultEnum.DATA_IS_NULL.getMsg());
+        } catch (Exception e) {
+            return Result.error(ResultEnum.DELETE_FAIL.getCode(), ResultEnum.DELETE_FAIL.getMsg());
+        }
+    }
+    @ApiOperation(value = "通过不同条件查询员工信息（支持模糊查询）")
+    @GetMapping("/getEmp")
+    public Result getEmp(Employee employee){
+        List<Map<String,Object>> res = employeeService.getEmp(employee);
+        if (res != null&&!res.isEmpty()) {
             return Result.success(res);
         } else {
             return Result.error(ResultEnum.DATA_IS_NULL.getCode(), ResultEnum.DATA_IS_NULL.getMsg());
         }
     }
-    @DeleteMapping("/deleteemployee")
-    public Result deleteEmployee(String id){
+    @DeleteMapping("/deleteBatch")
+    @ApiOperation(value = "通过员工编号批量删除员工")
+    public Result deleteBatch(@RequestParam Integer[] arr){
         try {
-            Employee res=employeeService.getEmployeeById(id);
-            if(res!=null){
-                employeeService.deleteEmployee(id);
-                return Result.success();
-            }
-            return Result.error(ResultEnum.DATA_IS_NULL.getCode(),ResultEnum.DATA_IS_NULL.getMsg());
-        } catch (Exception e) {
-            return Result.error(ResultEnum.DELETE_FAIL.getCode(),ResultEnum.DELETE_FAIL.getMsg());
+                int rows = employeeService.deleteBatch(arr);
+                if(rows>=1){
+                    return Result.success();
+                }
+                else{
+                    return Result.error(ResultEnum.DELETE_FAIL.getCode(), ResultEnum.DELETE_FAIL.getMsg());
+                }
+        } catch (Exception e){
+            return Result.error(ResultEnum.DELETE_FAIL.getCode(), ResultEnum.DELETE_FAIL.getMsg());
         }
     }
+
 }
